pyproject.toml
[project]
name = "travel-planner-assistant"
version = "0.1.0"
description = "WeCom travel planning assistant MVP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "pydantic>=2.10.0,<3.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "pycryptodome>=3.21.0,<4.0.0",
    "redis>=5.2.0,<6.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "websockets>=15.0.1,<16.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.25.0,<1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.providers", "app.services"]
